+++ src/modules/home/Home.vue
@@ -119,7 +119,7 @@ import CardOrganization from "./components/card-organization/CardOrganization.vu
 import EventCarousel from "./components/event-carousel/EventCarousel.vue";
 import BaseTransparentButton from "@/commons/UI/BaseTransparentButton.vue";
 import ArrowRightIcon from "@/assets/ArrowRight.vue";
-import { testData } from "./useHome/testData";
+import { testData } from "./use-home/testData";
 
 export default defineComponent({
   components: {
